Event rework
Entirely rework the event handling system. Now, an Event is allocated (from the
pool) for every event submission, and holds a pointer to an EventTarget. This
allows, for example, chained reads/sends into/from a buffer to correctly update
its indices as they complete. Additionally, it makes it possible to read all of
the CQEs from the CQ in one go, thus preventing EBUSY errors on SQE submission.