EVENT: Try to detect error conditions.
Previously, error conditions were managed solely by event handlers based
on the status returned from io_uring. However, this is not sufficient to
detect a short read or write, since it otherwise appears as a
"successful" return.

This patch introduces a mechanism for tracking the expected return value
of a particular event, whether that is a particular value, or range of
values. Events which show an unexpected return are flagged as failed,
which is passed on to event handlers for which it is relevant.

In the future, this will be used to handle short splices.

The flag system is getting somewhat overcomplicated, and there is
redundancy between the chain and ignore flags. This should be reworked
in the future.