HTTP: Split splice handling in response to a special function.
Handling for splice now goes through `http_response_splice_handle`, instead of directly to
`http_response_handle`, allowing (in the future) special handling. This will enable the tracking of
how much of the response body has been sent so far, allowing short writes to be handled correctly.
Before then, some mechanism must be devised to distinguish between splices from the file to the pipe
and from the pipe to the socket. This could potentially be done by an extra flag in the event.