INIT: Start building out some initial scaffolding
This begins take three of the grand C++ rewrite.

This initial patch builds out a whole bunch of initial scaffolding. Not
much actual functionality is here, aside from the beginnings of
coroutine tooling and an io_uring interface. The latter is only capable
of invoking OP_READ, and is not yet wired into the larger coroutine
system.