INIT: Start sketching some things out
This patch does some very basic scaffolding towards a simple echo
server. Many functions are not actually implemented. Only io_uring is
targeted, but the beginnings of a platform abstraction are provided.

Annoyingly, this cannot be built with GCC (it ICEs), and even Clang's
module implementation is broken enough that it requires random
inclusions of <climits> strewn around to avoid mismatched definitions
(realpath()).